base inicial implementada

+++ cardapio/db/schemas.py
@@ -4,7 +4,7 @@ from pydantic import BaseModel
 class ProdutoBase(BaseModel):
     nome: str
     descricao: str
-    preco: int
+    preco: float
     tipo: int
